refactoring: extract the show-compare-diff code to a new file src/consumer/component-ops/components-object-diff.js
